requirements.txt
accelerate==0.26.1
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.2
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.26.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
-e git+ssh://git@github.com/IDEA-Research/Grounded-Segment-Anything.git@fe24c033820adffff66ac0eb191828542e8afe5e#egg=groundingdino&subdirectory=GroundingDINO
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
json-tricks==3.17.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv-full==1.7.1
mmpose==0.28.0
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnxruntime==1.17.0
openai==1.11.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
-e git+https://github.com/xinyu1205/recognize-anything.git@b189bd1606694cf41110ed9be57d2e8674fd911d#egg=ram
regex==2023.12.25
requests==2.28.2
rich==13.4.2
safetensors==0.4.2
scipy==1.10.0
-e git+ssh://git@github.com/IDEA-Research/Grounded-Segment-Anything.git@fe24c033820adffff66ac0eb191828542e8afe5e#egg=segment_anything&subdirectory=segment_anything
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.65.2
traitlets==5.14.1
transformers==4.25.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.2.13
xtcocotools==1.14.3
yapf==0.40.2
zipp==3.17.0
